requirements.txt
clickhouse-driver==0.2.8
colorama==0.4.6
numpy==2.0.1
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1
tzlocal==5.2
